Dynamic EQ: up to 24 click-to-add bands, HP/LP slope, dynamics knee/times
Editor and DSP follow-ups on the pastel UI redesign.

- Bands: max 6 -> 24, all off by default. Double-click the curve to add a
  band (type from x position: HP | low-shelf | bell | high-shelf | LP;
  gain from y), double-click a node to remove it. Inactive bands skip
  per-block coefficient work and per-sample processing.
- HP/LP variable slope 12..96 dB/oct via a Butterworth biquad cascade
  (core/Filters.h designHpLpStage); Q adds resonance on the peak section.
  At 12 dB/oct this is identical to the previous single biquad.
- Dynamics: expose Attack/Release parameters and add a soft Knee
  (core/DynamicEqBand; dynamicOffsetDb gains a defaulted knee arg, knee=0
  reproduces the old hard-knee ramp exactly).
- Band panel: drop tabs, split into EQ (left) / Dynamics (right); add Freq
  and Gain knobs; Freq shows integer Hz, Q two decimals.
- Fix: selecting another band no longer overwrites the previously-selected
  band's parameters (reset APVTS attachments before recreating them).
- Analyzer FFT 2048 -> 8192 and ring 8192 -> 16384 so sub-100 Hz bins stay
  resolved at high sample rates (e.g. 192 kHz).
- Fix C4244 warning; set dark value-text colours directly on the sliders.

Tests: new z-domain oracles for the HP/LP cascade (-3 dB at cutoff for any
order, ~12n dB/oct, monotonic/maximally-flat, resonance) and the soft-knee
function; existing oracles unchanged. All 21 plugin tests pass and pluginval
strictness 5 (headless) passes locally.

+++ plugins/dynamic-eq/Source/FactoryLookAndFeel.h
@@ -131,7 +131,7 @@ public:
         g.setColour (text());
         g.setFont (juce::Font (juce::FontOptions (13.0f, juce::Font::bold)));
         g.drawText (b.getButtonText(),
-                    r.withTrimmedLeft ((int) (box.getWidth() + 8.0f)),
+                    r.withTrimmedLeft (box.getWidth() + 8.0f),
                     juce::Justification::centredLeft);
     }
